pyproject.toml
[tool.poetry]
name = "zakhar_core"
version = "0.1.0"
description = ""
authors = ["Andrei Gramakov <mail@agramakov.me>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.1"
smbus2 = "^0.3.0"
Adafruit_GPIO = "^1.0.3"
Adafruit_SSD1306 = "^1.6.2"
pillow = "^7.1.1"
"RPi.GPIO" = "^0.7.0"
numpy = "^1.18.2"
smbus = "^1.1.post2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.1"
rope = "^0.16.0"
yapf = "^0.30.0"
flake8 = "^3.7.9"
bandit = "^1.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
